+++ packages/react/src/dropdown/src/dropdown-portal.tsx
@@ -1,11 +1,22 @@
 import React from 'react';
 
-import { Popover, PopoverPortalProps } from '@melio-ui/popover';
+import { ContainerType, usePortal } from '@melio-ui/use-portal';
 
-export interface DropdownPortalProps extends PopoverPortalProps {}
+export interface DropdownPortalProps {
+  children?: React.ReactNode;
+  /**
+   * Popover 이 mount 되는 node (해당 node에는 position: relative 속성이 필요.)
+   * @default document.body
+   */
+  container?: ContainerType;
+}
 
 export default function DropdownPortal(props: DropdownPortalProps): JSX.Element {
-  return <Popover.Portal {...props} />;
+  const { container, children } = props;
+
+  const { Portal } = usePortal(container);
+
+  return <Portal>{children}</Portal>;
 }
 
 DropdownPortal.displayName = 'Dropdown.Portal';