+++ packages/react/src/popover/src/popover.stories.css.ts
@@ -10,3 +10,20 @@ export const content = style({
     outline: 0,
   },
 });
+
+export const arrow = style({
+  fill: 'var(--component-bg)',
+});
+
+export const close = style({
+  borderRadius: '100%',
+  height: '1.5625rem',
+  width: '1.5625rem',
+  display: 'inline-flex',
+  alignItems: 'center',
+  justifyContent: 'center',
+  // color: '#fff',
+  position: 'absolute',
+  top: '0.3125rem',
+  right: '0.3125rem',
+});