+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @melio-ui/react
 
+## 0.0.13
+
+### Patch Changes
+
+- 에러로 인해 다시 배포
+
+## 0.0.12
+
+### Patch Changes
+
+- popover/dropdown popper로 개선
+
 ## 0.0.11
 
 ### Patch Changes