+++ packages/react/src/popper/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @melio-ui/popper
 
+## 0.0.4
+
+### Patch Changes
+
+- 에러로 인해 다시 배포
+
+## 0.0.3
+
+### Patch Changes
+
+- popover/dropdown popper로 개선
+
 ## 0.0.1
 
 ### Patch Changes