+++ packages/react/src/tooltip/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @melio-ui/tooltip
 
+## 0.0.6
+
+### Patch Changes
+
+- 에러로 인해 다시 배포
+- Updated dependencies
+  - @melio-ui/popper@0.0.4
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @melio-ui/popper@0.0.3
+
 ## 0.0.4
 
 ### Patch Changes